+++ frontend/src/pages/xray/routing/types.ts
@@ -1,0 +1,16 @@
+export interface RuleRow {
+  key: number;
+  domain?: string;
+  ip?: string;
+  port?: string;
+  sourcePort?: string;
+  vlessRoute?: string;
+  network?: string;
+  sourceIP?: string;
+  user?: string;
+  inboundTag?: string;
+  protocol?: string;
+  attrs?: string;
+  outboundTag?: string;
+  balancerTag?: string;
+}
